requirements.txt
earthengine-api
functions-framework
geopandas
google-cloud-storage
pandas
python-dotenv
requests